orchestration/pyproject.toml
[project]
name = "orchestration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "paho-mqtt>=2.1.0",
]
